TASK_MANAGEMENT_REFACTORED/backend/requirements.txt
agno==2.0.8
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.2
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
channels==4.3.1
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cron_descriptor==2.0.6
cryptography==46.0.1
cssselect==1.3.0
daphne==4.2.1
distro==1.9.0
dj-database-url==3.0.1
Django==5.2.6
django-auditlog==3.3.0
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-countries==7.6.1
django-easy-audit==1.3.7
django-encrypted-model-fields==0.6.5
django-jazzmin==3.0.1
django-jsignature==0.12
django-timezone-field==7.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docstring_parser==0.17.0
drf-spectacular==0.28.0
geographiclib==2.1
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.45
groq==0.31.1
h11==0.16.0
holidays==0.81
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
lxml==6.0.2
markdown-it-py==4.0.0
mdurl==0.1.2
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.52
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
pyquery==2.0.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
redis==6.4.0
referencing==0.36.2
requests==2.32.5
rich==14.1.0
rpds-py==0.27.1
service-identity==24.2.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
slugify==0.0.1
smmap==5.0.2
sniffio==1.3.1
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.6.1
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.11.0
zope.interface==8.0
